+++ chat-widget/src/components/webchatcontainerstateful/common/DesignerChatAdapter.ts
@@ -1,7 +1,7 @@
 import { Message } from "botframework-directlinejs";
 import { Observable } from "rxjs/Observable";
 import MockAdapter from "./mockadapter";
-import { customerUser, postBotMessageActivity, postEchoActivity, postSystemMessageActivity } from "./utils/chatAdapterUtils";
+import { customerUser, postAgentMessageActivity, postBotMessageActivity, postEchoActivity, postSystemMessageActivity } from "./utils/chatAdapterUtils";
 
 export class DesignerChatAdapter extends MockAdapter {
     constructor() {
@@ -9,10 +9,12 @@ export class DesignerChatAdapter extends MockAdapter {
 
         setTimeout(() => {
             postBotMessageActivity(this.activityObserver, "Thank you for contacting us! How can I help you today?", undefined, 0);
+            postBotMessageActivity(this.activityObserver, "Please accept terms and conditions to proceed. Visit the link for terms and conditions <a href=\"\">here</a>.", undefined, 0);
             this.postUserActivity("I need to change my address.", 0);
             postBotMessageActivity(this.activityObserver, "Okay, let me connect you with a live agent.", undefined, 100);
             postSystemMessageActivity(this.activityObserver, "John has joined the chat", 100);
-            postBotMessageActivity(this.activityObserver, "I'd be happy to help you update your account.", undefined, 100);
+            postAgentMessageActivity(this.activityObserver, "I'd be happy to help you update your account.", undefined, 100);
+            this.postUserActivity("I have trouble visiting the signin page <a href=\"\">signin</a>.", 0);
         }, 1000);
     }